implement output adapter of place search domain
