+++ subprojects/bounded_context/place_search/build.gradle.kts
@@ -1,2 +1,4 @@
 dependencies {
+    outputAdapterImplementation(project(":bounded_context:place", "application"))
+    outputAdapterImplementation(project(":bounded_context:accessibility", "application"))
 }
