+++ subprojects/bounded_context/place_search/src/domain/kotlin/club/staircrusher/place_search/domain/model/PlaceAccessibility.kt
@@ -1,0 +1,4 @@
+package club.staircrusher.place_search.domain.model
+
+class PlaceAccessibility {
+}
