+++ subprojects/bounded_context/place_search/src/domain/kotlin/club/staircrusher/place_search/domain/service/AccessibilityService.kt
@@ -1,0 +1,9 @@
+package club.staircrusher.place_search.domain.service
+
+import club.staircrusher.place_search.domain.model.BuildingAccessibility
+import club.staircrusher.place_search.domain.model.Place
+import club.staircrusher.place_search.domain.model.PlaceAccessibility
+
+interface AccessibilityService {
+    fun getAccessibility(place: Place): Pair<PlaceAccessibility?, BuildingAccessibility?>
+}
