+++ subprojects/bounded_context/place_search/src/domain/kotlin/club/staircrusher/place_search/domain/service/PlaceService.kt
@@ -1,0 +1,7 @@
+package club.staircrusher.place_search.domain.service
+
+import club.staircrusher.place_search.domain.model.Place
+
+interface PlaceService {
+    suspend fun findByKeyword(keyword: String): List<Place>
+}
