--- subprojects/bounded_context/place_search/src/domain/kotlin/club/staircrusher/place_search/model/BuildingAccessibility.kt
@@ -1,4 +1,0 @@
-package club.staircrusher.place_search.model
-
-class BuildingAccessibility {
-}
