--- subprojects/bounded_context/place_search/src/domain/kotlin/club/staircrusher/place_search/service/AccessibilityService.kt
@@ -1,9 +1,0 @@
-package club.staircrusher.place_search.service
-
-import club.staircrusher.place_search.model.BuildingAccessibility
-import club.staircrusher.place_search.model.Place
-import club.staircrusher.place_search.model.PlaceAccessibility
-
-interface AccessibilityService {
-    fun getAccessibility(place: Place): Pair<PlaceAccessibility?, BuildingAccessibility?>
-}
