--- subprojects/bounded_context/place_search/src/domain/kotlin/club/staircrusher/place_search/service/PlaceService.kt
@@ -1,7 +1,0 @@
-package club.staircrusher.place_search.service
-
-import club.staircrusher.place_search.model.Place
-
-interface PlaceService {
-    suspend fun findByKeyword(keyword: String): List<Place>
-}
