--- subprojects/bounded_context/place_search/src/output-adapter/kotlin/club/staircrusher/place_search/service/InMemoryAccessibilityService.kt
@@ -1,11 +1,0 @@
-package club.staircrusher.place_search.service
-
-import club.staircrusher.place_search.model.BuildingAccessibility
-import club.staircrusher.place_search.model.Place
-import club.staircrusher.place_search.model.PlaceAccessibility
-
-class InMemoryAccessibilityService : AccessibilityService {
-    override fun getAccessibility(place: Place): Pair<PlaceAccessibility?, BuildingAccessibility?> {
-        TODO("Not yet implemented")
-    }
-}
