--- subprojects/bounded_context/place_search/src/output-adapter/kotlin/club/staircrusher/place_search/service/InMemoryPlaceService.kt
@@ -1,9 +1,0 @@
-package club.staircrusher.place_search.service
-
-import club.staircrusher.place_search.model.Place
-
-class InMemoryPlaceService : PlaceService {
-    override suspend fun findByKeyword(keyword: String): List<Place> {
-        TODO("Not yet implemented")
-    }
-}
